# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Contacts API",
      "description": "API to allow CRUD operations in Mongodb Database",
      "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
      "https"
    ],
    "paths": {
      "/api-docs": {
        "get": {
          "description": "Retrieves api docs data",
          "parameters": [],
          "responses": {}
        }
      },
      "/contacts/": {
        "get": {
          "description": "Retrieves all contacts from the specified collection",
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "post": {
          "description": "Adding a new contact to the database",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "favoriteColor": {
                    "example": "any"
                  },
                  "birthday": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/contacts/{id}": {
        "get": {
          "description": "Retrieves contact with the specified ID",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "put": {
          "description": "Updates an existing contact as specified by the ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "favoriteColor": {
                    "example": "any"
                  },
                  "birthday": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "description": "Deletes the specified ID from the database",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    }
  }